+++ packages/devextreme/js/__internal/scheduler/__tests__/workspace.test.ts
@@ -1,0 +1,198 @@
+import {
+  afterEach, beforeEach, describe, expect, it, jest,
+} from '@jest/globals';
+import $ from '@js/core/renderer';
+import { getWidth } from '@js/core/utils/size';
+
+import fx from '../../../common/core/animation/fx';
+import CustomStore from '../../../data/custom_store';
+import { getResourceManagerMock } from '../__mock__/resource_manager.mock';
+import SchedulerTimelineDay from '../workspaces/m_timeline_day';
+import SchedulerTimelineMonth from '../workspaces/m_timeline_month';
+import SchedulerTimelineWeek from '../workspaces/m_timeline_week';
+import SchedulerTimelineWorkWeek from '../workspaces/m_timeline_work_week';
+import type SchedulerWorkSpace from '../workspaces/m_work_space';
+import SchedulerWorkSpaceDay from '../workspaces/m_work_space_day';
+import SchedulerWorkSpaceMonth from '../workspaces/m_work_space_month';
+import SchedulerWorkSpaceWeek from '../workspaces/m_work_space_week';
+import SchedulerWorkSpaceWorkWeek from '../workspaces/m_work_space_work_week';
+import { createScheduler } from './__mock__/create_scheduler';
+import { setupSchedulerTestEnvironment } from './__mock__/m_mock_scheduler';
+
+const CLASSES = {
+  scheduler: 'dx-scheduler',
+  workSpace: 'dx-scheduler-work-space',
+};
+
+type WorkspaceConstructor<T> = new (container: Element, options?: any) => T;
+
+const createWorkspace = <T extends SchedulerWorkSpace>(
+  WorkSpace: WorkspaceConstructor<T>,
+  currentView: string,
+): T => {
+  const container = document.createElement('div');
+  const workspace = new WorkSpace(container, {
+    views: [currentView],
+    currentView,
+    currentDate: new Date(2017, 4, 25),
+    firstDayOfWeek: 0,
+    getResourceManager: () => getResourceManagerMock([]),
+  });
+  (workspace as any)._isVisible = () => true;
+  expect(container.classList).toContain('dx-scheduler-work-space');
+
+  return workspace;
+};
+
+const workSpaces: {
+  currentView: string;
+  WorkSpace: WorkspaceConstructor<SchedulerWorkSpace>;
+}[] = [
+  { currentView: 'day', WorkSpace: SchedulerWorkSpaceDay },
+  { currentView: 'week', WorkSpace: SchedulerWorkSpaceWeek },
+  { currentView: 'workWeek', WorkSpace: SchedulerWorkSpaceWorkWeek },
+  { currentView: 'month', WorkSpace: SchedulerWorkSpaceMonth },
+  { currentView: 'timelineDay', WorkSpace: SchedulerTimelineDay },
+  { currentView: 'timelineWeek', WorkSpace: SchedulerTimelineWeek },
+  { currentView: 'timelineWorkWeek', WorkSpace: SchedulerTimelineWorkWeek },
+  { currentView: 'timelineMonth', WorkSpace: SchedulerTimelineMonth },
+];
+
+describe('Workspace', () => {
+  describe('Base functionality', () => {
+    workSpaces.forEach(({ currentView, WorkSpace }) => {
+      it(`should clear cache on dimension change, view: ${currentView}`, () => {
+        const workspace = createWorkspace(WorkSpace, currentView);
+        jest.spyOn(workspace.cache, 'clear');
+
+        workspace.cache.memo('test', () => 'value');
+        workspace._dimensionChanged();
+
+        expect(workspace.cache.clear).toHaveBeenCalledTimes(1);
+      });
+
+      it(`should clear cache on _cleanView call, view: ${currentView}`, () => {
+        const workspace = createWorkspace(WorkSpace, currentView);
+        jest.spyOn(workspace.cache, 'clear');
+
+        workspace.cache.memo('test', () => 'value');
+        workspace._cleanView();
+
+        expect(workspace.cache.clear).toHaveBeenCalledTimes(1);
+        expect(workspace.cache.size).toBe(0);
+      });
+    });
+  });
+
+  describe('Recalculation with Async Templates', () => {
+    beforeEach(() => {
+      fx.off = true;
+      setupSchedulerTestEnvironment({ height: 600 });
+    });
+
+    afterEach(() => {
+      const $scheduler = $(document.querySelector(`.${CLASSES.scheduler}`));
+      // @ts-expect-error
+      $scheduler.dxScheduler('dispose');
+      document.body.innerHTML = '';
+      fx.off = false;
+    });
+
+    it('should not duplicate workspace elements when resources are loaded asynchronously (T661335)', async () => {
+      const { scheduler, container } = await createScheduler({
+        templatesRenderAsynchronously: true,
+        currentView: 'day',
+        views: ['day'],
+        groups: ['owner'],
+        resources: [
+          {
+            fieldExpr: 'owner',
+            dataSource: [{ id: 1, text: 'Owner 1' }],
+          },
+          {
+            fieldExpr: 'room',
+            dataSource: new CustomStore({
+              load(): Promise<unknown> {
+                return new Promise((resolve) => {
+                  setTimeout(() => {
+                    resolve([{ id: 1, text: 'Room 1', color: '#ff0000' }]);
+                  });
+                });
+              },
+            }),
+          },
+        ],
+        dataSource: [
+          {
+            text: 'Meeting in Room 1',
+            startDate: new Date(2017, 4, 25, 9, 0),
+            endDate: new Date(2017, 4, 25, 10, 0),
+            roomId: 1,
+          },
+        ],
+        startDayHour: 9,
+        currentDate: new Date(2017, 4, 25),
+        height: 600,
+      });
+
+      scheduler.option('groups', ['room']);
+
+      await new Promise((r) => { setTimeout(r); });
+
+      const $workSpaces = $(container).find(`.${CLASSES.workSpace}`);
+      const $groupHeader = $(container).find('.dx-scheduler-group-header');
+
+      expect($workSpaces.length).toBe(1);
+
+      expect($groupHeader.length).toBeGreaterThan(0);
+      expect($groupHeader.text()).toContain('Room 1');
+    });
+  });
+
+  describe('scrollTo', () => {
+    beforeEach(() => {
+      fx.off = true;
+      setupSchedulerTestEnvironment({ height: 600 });
+    });
+
+    afterEach(() => {
+      document.body.innerHTML = '';
+      fx.off = false;
+    });
+
+    it('should scroll to date with offset (T1310544)', async () => {
+      const { scheduler } = await createScheduler({
+        views: ['timelineDay'],
+        currentView: 'timelineDay',
+        currentDate: new Date(2021, 1, 2),
+        firstDayOfWeek: 0,
+        startDayHour: 6,
+        endDayHour: 18,
+        offset: 720,
+        cellDuration: 60,
+        height: 580,
+      });
+
+      const workspace = scheduler.getWorkSpace();
+      const scrollable = workspace.getScrollable();
+      const $scrollable = scrollable.$element();
+      const scrollBySpy = jest.spyOn(scrollable, 'scrollBy');
+
+      // With offset: 720 (12 hours), cells start at 18:00 (6:00 + 12h)
+      // For date 22:00, this should be cell index 4 (18:00=0, 19:00=1, 20:00=2, 21:00=3, 22:00=4)
+      const leftCellCount = 4;
+      const cellWidth = workspace.getCellWidth();
+      const scrollableWidth = getWidth($scrollable);
+      const expectedLeft = leftCellCount * cellWidth - (scrollableWidth - cellWidth) / 2;
+
+      const targetDate = new Date(2021, 1, 2, 22, 0);
+      scheduler.scrollTo(targetDate, undefined, false);
+
+      expect(scrollBySpy).toHaveBeenCalledTimes(1);
+      const scrollParams = scrollBySpy.mock.calls[0][0] as { left: number; top: number };
+      expect(scrollParams.left).toBeCloseTo(expectedLeft, 1);
+
+      scrollBySpy.mockRestore();
+    });
+  });
+});
